+++ packages/@coorpacademy-components/src/organism/review-no-skills/prop-types.ts
@@ -1,0 +1,5 @@
+export type NoSkillsProps = {
+  titleNoSkills?: string;
+  textNoSkills?: string;
+  iconSkillAriaLabel?: string;
+};
