+++ packages/@coorpacademy-components/src/organism/review-no-skills/test/fixtures/no-skills.js
@@ -1,0 +1,8 @@
+export default {
+  props: {
+    titleNoSkills: 'You have no skill to revise yet',
+    textNoSkills:
+      'Massa molestie ut sed malesuada. Cursus aliquet enim scelerisque sagittis at rutrum aliquam.',
+    iconSkillAriaLabel: 'Image without information'
+  }
+};
