+++ packages/@coorpacademy-components/src/organism/review-skills/prop-types.ts
@@ -5,39 +5,6 @@ export const ListSkillsPropTypes = {
   listSkills: PropTypes.arrayOf(PropTypes.shape(SkillCardPropsType))
 };
 
-export const NoSkillsPropTypes = {
-  titleNoSkills: PropTypes.string,
-  textNoSkills: PropTypes.string,
-  iconSkillAriaLabel: PropTypes.string
-};
-
-const propTypes = {
-  'aria-label': PropTypes.string,
-  title: PropTypes.string,
-  isLoading: PropTypes.bool,
-  isLoadingAriaLabel: PropTypes.string,
-  listSkills: PropTypes.arrayOf(PropTypes.shape(SkillCardPropsType)),
-  titleNoSkills: PropTypes.string,
-  textNoSkills: PropTypes.string,
-  iconSkillAriaLabel: PropTypes.string
-};
-
-export default propTypes;
-
 export type ItemProps = {
   item: SkillCardProps;
-};
-
-export type NoSkillsProps = {
-  titleNoSkills?: string;
-  textNoSkills?: string;
-  iconSkillAriaLabel?: string;
-};
-
-export type ReviewSkillsProps = NoSkillsProps & {
-  'aria-label': string;
-  title: string;
-  isLoading: boolean;
-  isLoadingAriaLabel: string;
-  listSkills: SkillCardProps[];
 };
