+++ packages/@coorpacademy-components/src/organism/review-skills/test/fixtures/many-skills.js
@@ -1,32 +1,26 @@
 import cardDefault from '../../../../molecule/skill-card/test/fixtures/default';
 
-const listSkills = [
-  {...cardDefault.props, skillTitle: 'Digital communication', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Creativity and innovation', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Agility', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Sustainable thinking', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Critical thinking and analysis', isCustom: true},
-  {...cardDefault.props, skillTitle: 'Negotiation and powers of persuasion', isCustom: true},
-  {...cardDefault.props, skillTitle: 'Digital communication', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Creativity and innovation', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Agility', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Sustainable thinking', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Critical thinking and analysis', isCustom: true},
-  {...cardDefault.props, skillTitle: 'Negotiation and powers of persuasion', isCustom: true},
-  {...cardDefault.props, skillTitle: 'Digital communication', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Creativity and innovation', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Agility', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Sustainable thinking', isCustom: false},
-  {...cardDefault.props, skillTitle: 'Critical thinking and analysis', isCustom: true},
-  {...cardDefault.props, skillTitle: 'Negotiation and powers of persuasion', isCustom: true}
-];
-
 export default {
   props: {
-    'aria-label': 'Review Skills Container',
-    title: 'Skills you can revise',
-    listSkills,
-    isLoading: false,
-    isLoadingAriaLabel: 'Review skills container is loading'
+    listSkills: [
+      {...cardDefault.props, skillTitle: 'Digital communication', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Creativity and innovation', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Agility', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Sustainable thinking', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Critical thinking and analysis', isCustom: true},
+      {...cardDefault.props, skillTitle: 'Negotiation and powers of persuasion', isCustom: true},
+      {...cardDefault.props, skillTitle: 'Digital communication', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Creativity and innovation', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Agility', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Sustainable thinking', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Critical thinking and analysis', isCustom: true},
+      {...cardDefault.props, skillTitle: 'Negotiation and powers of persuasion', isCustom: true},
+      {...cardDefault.props, skillTitle: 'Digital communication', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Creativity and innovation', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Agility', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Sustainable thinking', isCustom: false},
+      {...cardDefault.props, skillTitle: 'Critical thinking and analysis', isCustom: true},
+      {...cardDefault.props, skillTitle: 'Negotiation and powers of persuasion', isCustom: true}
+    ]
   }
 };
